streams_logger_test/requirements.txt
kafka-python==2.0.2
aiokafka